refactor: move bounded_sample and create_run_metadata out of persistence
Move bounded_sample to shared/sampling.py (general iterator utility).
Move create_run_metadata to use_cases/run/sample_builder.py (its only
production consumer).

Remove both functions from adapters/persistence/runlog.py and update
all importers (1 production + 7 test files).

Closes #748
