+++ apps/server/tests/integration/test_runtime_quality_pass_regressions.py
@@ -20,9 +20,9 @@ import numpy as np
 import pytest
 
 from vibesensor.adapters.persistence.history_db import HistoryDB
-from vibesensor.adapters.persistence.runlog import bounded_sample as _bounded_sample
 from vibesensor.infra.config.settings_store import SettingsStore
 from vibesensor.infra.processing import SignalProcessor
+from vibesensor.shared.sampling import bounded_sample as _bounded_sample
 
 # -- shared helpers ----------------------------------------------------------
